ENH Rename to miin_dependencies
